+++ internal/amd64/numeric_avx2_amd64.go
@@ -121,10 +121,4 @@ func scatterUint64AVX2(dst []uint64, idx []int32, src []uint64)
 func movingAverageFloat32AVX2(dst []float32, a []float32, width int)
 
 //go:noescape
-func matMulFloat32AVX2(dst []float32, a []float32, b []float32, m int, k int, n int)
-
-//go:noescape
 func movingAverageFloat64AVX2(dst []float64, a []float64, width int)
-
-//go:noescape
-func matMulFloat64AVX2(dst []float64, a []float64, b []float64, m int, k int, n int)