+++ internal/amd64/numeric_sse2_amd64.go
@@ -85,10 +85,4 @@ func tileUint64SSE2(dst []uint64, pattern []uint64)
 func movingAverageFloat32SSE2(dst []float32, a []float32, width int)
 
 //go:noescape
-func matMulFloat32SSE2(dst []float32, a []float32, b []float32, m int, k int, n int)
-
-//go:noescape
 func movingAverageFloat64SSE2(dst []float64, a []float64, width int)
-
-//go:noescape
-func matMulFloat64SSE2(dst []float64, a []float64, b []float64, m int, k int, n int)