+++ internal/arm64/numeric_sve2_arm64.go
@@ -94,10 +94,4 @@ func tileUint64SVE2(dst []uint64, pattern []uint64)
 func movingAverageFloat32SVE2(dst []float32, a []float32, width int)
 
 //go:noescape
-func matMulFloat32SVE2(dst []float32, a []float32, b []float32, m int, k int, n int)
-
-//go:noescape
 func movingAverageFloat64SVE2(dst []float64, a []float64, width int)
-
-//go:noescape
-func matMulFloat64SVE2(dst []float64, a []float64, b []float64, m int, k int, n int)